+++ packages/integration-test/simple.test.mjs
@@ -1,26 +1,28 @@
 /** @jest-environment jsdom */
 
-import { act, renderHook } from '@testing-library/react';
+import { act } from '@testing-library/react';
 import { createPropagation } from 'use-propagate';
 
+const renderHook =
+  // eslint-disable-next-line @typescript-eslint/no-var-requires
+  require('@testing-library/react').renderHook ||
+  // eslint-disable-next-line @typescript-eslint/no-var-requires
+  require('@testing-library/react-hooks').renderHook;
+
 test('When a value is being propagated, listener should receive the value', () => {
-  const { Provider, useListen, usePropagate } = createPropagation();
+  const { useListen, usePropagate } = createPropagation();
   const listener = jest.fn();
+  let propagate;
 
-  const result = renderHook(
-    ({ value }) => {
-      const propagate = usePropagate();
+  renderHook(() => {
+    propagate = usePropagate();
 
-      useListen(listener);
-
-      typeof value === 'undefined' || propagate(value);
-    },
-    { initialProps: {}, wrapper: Provider }
-  );
+    useListen(listener);
+  });
 
   expect(listener).toHaveBeenCalledTimes(0);
 
-  act(() => result.rerender({ value: 'Hello, World!' }));
+  act(() => propagate('Hello, World!'));
 
   expect(listener).toHaveBeenCalledTimes(1);
   expect(listener).toHaveBeenNthCalledWith(1, 'Hello, World!');