+++ packages/use-propagate/src/createPropagation.tsx
@@ -1,30 +1,53 @@
-import { useEffect } from 'react';
+import { useCallback, useEffect, useLayoutEffect, useMemo } from 'react';
 import { useRefFrom } from 'use-ref-from';
-
-import removeInline from './private/removeInline';
 
 type Listener<T> = (value: T) => void;
 
-export default function createPropagation<T>() {
+type Init = {
+  /**
+   * `true` to allows calling propagate callback function during render-time, otherwise, `false` (default).
+   *
+   * Propagation during render-time is normally discouraged. If listeners save the value into a state, multiple re-render could occur.
+   * This option prevents render deadlock by disallowing propagation during render-time.
+   */
+  allowPropagateDuringRender?: boolean | undefined;
+};
+
+export default function createPropagation<T>(init: Init = {}) {
   type Fn = Listener<T>;
 
-  const listeners: Fn[] = [];
+  let rendering: boolean = false;
 
-  const addListener = (listener: Fn): void => void listeners.push(listener);
-  const removeListener = (listener: Fn): void => removeInline(listeners, listener);
-  const usePropagate = () => (value: T) => listeners.forEach(listener => listener(value));
+  const { allowPropagateDuringRender } = init;
+  const listeners: Set<Fn> = new Set();
+
+  const addListener = (listener: Fn): void => void listeners.add(listener);
+  const removeListener = (listener: Fn): void => void listeners.delete(listener);
+  const usePropagate = () => {
+    rendering = true;
+
+    useLayoutEffect(() => {
+      rendering = false;
+    });
+
+    return (value: T) => {
+      if (rendering && !allowPropagateDuringRender) {
+        return console.warn(
+          'use-propagate: The propagate callback function should not be called while rendering, ignoring the call.'
+        );
+      }
+
+      listeners.forEach(listener => listener(value));
+    };
+  };
 
   return {
     useListen: (listener: Fn) => {
       const listenerRef = useRefFrom(listener);
+      const wrappingListener = useCallback((value: T) => listenerRef.current(value), [listenerRef]);
 
-      useEffect(() => {
-        const wrappingListener = (value: T) => listenerRef.current(value);
-
-        addListener(wrappingListener);
-
-        return () => removeListener(wrappingListener);
-      }, [listenerRef]);
+      useMemo(() => addListener(wrappingListener), [wrappingListener]);
+      useEffect(() => () => removeListener(wrappingListener), [wrappingListener]);
     },
     usePropagate
   };